+++ angel_container/lib/angel_container.dart
@@ -1,3 +1,4 @@
+export 'src/empty/empty.dart';
 export 'src/container.dart';
 export 'src/exception.dart';
 export 'src/reflector.dart';